Convert GRAVCore into a dll to allow for debug mode because of libtorch

+++ core/src/application.cpp
@@ -92,7 +92,7 @@ void GRAVEngine::application::run()
 
 		// Get the delta time
 		float time = (float)glfwGetTime();
-		Time::timestep timestep = Time::timestep(time - m_LastFrameTime);
+		Time::timestep timestep = Time::timestep((float)(time - m_LastFrameTime));
 		m_LastFrameTime = time;
 
 		if (m_Minimized == false)
